refactor: move runner out of server, the server uses the runner
