+++ modules/importer/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod endpoints;
 pub mod model;
+pub mod runner;
 pub mod server;
 pub mod service;
 pub mod test;
