+++ modules/importer/src/runner/csaf/report.rs
@@ -1,6 +1,6 @@
-use crate::server::context::RunContext;
-use crate::server::{
+use crate::runner::{
     common::storage::StorageError,
+    context::RunContext,
     csaf::storage::StorageVisitor,
     report::{Phase, ReportVisitor},
 };
