+++ modules/importer/src/runner/sbom/report.rs
@@ -1,6 +1,6 @@
-use crate::server::context::RunContext;
-use crate::server::{
+use crate::runner::{
     common::storage::StorageError,
+    context::RunContext,
     report::{Phase, ReportVisitor},
     sbom::storage::StorageVisitor,
 };
